Weight changed to int?

+++ Katena/Domain/Entities/AnswersBase.cs
@@ -16,102 +16,102 @@ namespace Katena.Domain.Entities
 		//Вес
 		[Required]
         [Display(Name ="Принять решение(Производитель)")]
-        public double? Weight { get; set; }
+        public int? Weight { get; set; }
 
 		//Вес
 		[Required]
 		[Display(Name = "Фокус внимания(Производитель)")]
-		public double? Weight1 { get; set; }
+		public int? Weight1 { get; set; }
 
 		//Вес
 		[Required]
 		[Display(Name = "Процесс принятия решений(Производитель)")]
-		public double? Weight2 { get; set; }
+		public int? Weight2 { get; set; }
 
 		//Вес
 		[Required]
 		[Display(Name = "Ориентирован на(Производитель)")]
-		public double? Weight3 { get; set; }
+		public int? Weight3 { get; set; }
 
 		//Вес
 		[Required]
 		[Display(Name = "Принять решение(Администратор)")]
-		public double? Weight4 { get; set; }
+		public int? Weight4 { get; set; }
 
 		//Вес
 		[Required]
 		[Display(Name = "Фокус внимания(Администратор)")]
-		public double? Weight5 { get; set; }
+		public int? Weight5 { get; set; }
 
 		//Вес
 		[Required]
 		[Display(Name = "Процесс принятия решений(Администратор)")]
-		public double? Weight6 { get; set; }
+		public int? Weight6 { get; set; }
 
 		//Вес
 		[Required]
 		[Display(Name = "Ориентирован на(Администратор)")]
-		public double? Weight7 { get; set; }
+		public int? Weight7 { get; set; }
 
 		//Вес
 		[Required]
 		[Display(Name = "Принять решение(Предприниматель)")]
-		public double? Weight8 { get; set; }
+		public int? Weight8 { get; set; }
 
 		//Вес
 		[Required]
 		[Display(Name = "Фокус внимания(Предприниматель)")]
-		public double? Weight9 { get; set; }
+		public int? Weight9 { get; set; }
 
 		//Вес
 		[Required]
 		[Display(Name = "Процесс принятия решений(Предприниматель)")]
-		public double? Weight10 { get; set; }
+		public int? Weight10 { get; set; }
 
 		//Вес
 		[Required]
 		[Display(Name = "Ориентирован на(Предприниматель)")]
-		public double? Weight11 { get; set; }
+		public int? Weight11 { get; set; }
 
 		//Вес
 		[Required]
 		[Display(Name = "Принять решение(Интегратор)")]
-		public double? Weight12 { get; set; }
+		public int? Weight12 { get; set; }
 
 		//Вес
 		[Required]
 		[Display(Name = "Фокус внимания(Интегратор)")]
-		public double? Weight13 { get; set; }
+		public int? Weight13 { get; set; }
 
 		//Вес
 		[Required]
 		[Display(Name = "Процесс принятия решений(Интегратор)")]
-		public double? Weight14 { get; set; }
+		public int? Weight14 { get; set; }
 
 		//Вес
 		[Required]
 		[Display(Name = "Ориентирован на(Интегратор)")]
-		public double? Weight15 { get; set; }
+		public int? Weight15 { get; set; }
 
 		//Вес сложности варианта ответа
 		[Required]
 		[Display(Name= "Вес производитель")]
-		public double? typeWeight1 { get; set; }
+		public int? typeWeight1 { get; set; }
 
 		//Вес сложности варианта ответа
 		[Required]
 		[Display(Name = "Вес администартор")]
-		public double? typeWeight2 { get; set; }
+		public int? typeWeight2 { get; set; }
 		
 		//Вес сложности варианта ответа
 		[Required]
 		[Display(Name = "Вес предприниматель")]
-		public double? typeWeight3 { get; set; }
+		public int? typeWeight3 { get; set; }
 
 		//Вес сложности варианта ответа
 		[Required]
 		[Display(Name = "Вес интегратор")]
-		public double? typeWeight4 { get; set; }
+		public int? typeWeight4 { get; set; }
 
 		public List<Guid> reasonsId { get; set; }
 	}
