+++ Katena/Domain/Entities/ReasonBase.cs
@@ -13,75 +13,75 @@ namespace Katena.Domain.Entities
 		//Веса
 		[Required]
 		[Display(Name = "Поиск виноватого(З)")]
-		public double? Weight { get; set; }
+		public int? Weight { get; set; }
 
 		[Required]
 		[Display(Name = "Поиск ценностей(З)")]
-		public double? Weight1 { get; set; }
+		public int? Weight1 { get; set; }
 
 		[Required]
 		[Display(Name = "Вектор внимания(З)")]
-		public double? Weight2 { get; set; }
+		public int? Weight2 { get; set; }
 
 		[Required]
 		[Display(Name = "Проявление заботы(З)")]
-		public double? Weight3 { get; set; }
+		public int? Weight3 { get; set; }
 
 		[Required]
 		[Display(Name = "Ориентация во времени(З)")]
-		public double? Weight4 { get; set; }
+		public int? Weight4 { get; set; }
 
 		[Required]
 		[Display(Name = "Поиск виноватого(П)")]
-		public double? Weight5 { get; set; }
+		public int? Weight5 { get; set; }
 
 		[Required]
 		[Display(Name = "Поиск ценностей(П)")]
-		public double? Weight6 { get; set; }
+		public int? Weight6 { get; set; }
 
 		[Required]
 		[Display(Name = "Вектор внимания(П)")]
-		public double? Weight7 { get; set; }
+		public int? Weight7 { get; set; }
 
 		[Required]
 		[Display(Name = "Проявление заботы(П)")]
-		public double? Weight8 { get; set; }
+		public int? Weight8 { get; set; }
 
 		[Required]
 		[Display(Name = "Ориентация во времени(П)")]
-		public double? Weight9 { get; set; }
+		public int? Weight9 { get; set; }
 
 		[Required]
 		[Display(Name = "Поиск виноватого(Х)")]
-		public double? Weight10 { get; set; }
+		public int? Weight10 { get; set; }
 
 		[Required]
 		[Display(Name = "Поиск ценностей(Х)")]
-		public double? Weight11 { get; set; }
+		public int? Weight11 { get; set; }
 
         [Required]
         [Display(Name = "Вектор внимания(Х)")]
-        public double? Weight12 { get; set; }
+        public int? Weight12 { get; set; }
 
         [Required]
         [Display(Name = "Проявление заботы(Х)")]
-        public double? Weight13 { get; set; }
+        public int? Weight13 { get; set; }
 
         [Required]
         [Display(Name = "Ориентация во времени(Х)")]
-        public double? Weight14 { get; set; }
+        public int? Weight14 { get; set; }
 
 		//Веса сложности варианта мотивации
 		[Required]
 		[Display(Name = "Вес наставкник")]
-		public double? typeWeight1 { get; set; }
+		public int? typeWeight1 { get; set; }
 
 		[Required]
 		[Display(Name = "Вес ищущий")]
-		public double? typeWeight2 { get; set; }
+		public int? typeWeight2 { get; set; }
 
 		[Required]
 		[Display(Name = "Вес охотник")]
-		public double? typeWeight3 { get; set; }
+		public int? typeWeight3 { get; set; }
 	}
 }
